fix: change format-cpf to correct dir
