+++ src/lib/format-cpf/types.ts
@@ -1,0 +1,1 @@
+export type IFormatCpf = (cpf: string) => string
