--- types.ts
@@ -1,1 +1,0 @@
-export type IFormatCpf = (cpf: string) => string
